Setup inputs to update configurations
