--- media/main.js
@@ -1,15 +1,0 @@
-const vscode = acquireVsCodeApi();
-
-window.addEventListener("load", main);
-
-function main() {
-  const howdyButton = document.getElementById("howdy");
-  howdyButton.addEventListener("click", handleHowdyClick);
-}
-
-function handleHowdyClick() {
-  vscode.postMessage({
-    command: "hello",
-    text: "Hey there partner! 🤠",
-  });
-}
