+++ package.json
@@ -147,21 +147,21 @@
 		"test": "node ./out/test/runTest.js"
 	},
 	"devDependencies": {
-		"@types/vscode": "^1.60.0",
 		"@types/glob": "^7.1.3",
 		"@types/mocha": "^8.2.2",
-		"@types/node": "14.x",
+		"@types/node": "^14.17.26",
+		"@types/vscode": "^1.61.0",
+		"@typescript-eslint/eslint-plugin": "^4.33.0",
+		"@typescript-eslint/parser": "^4.33.0",
 		"eslint": "^7.27.0",
-		"@typescript-eslint/eslint-plugin": "^4.26.0",
-		"@typescript-eslint/parser": "^4.26.0",
-		"glob": "^7.1.7",
+		"glob": "^7.2.0",
 		"mocha": "^8.4.0",
-		"typescript": "^4.3.2",
+		"typescript": "^4.4.4",
 		"vscode-test": "^1.5.2"
 	},
 	"dependencies": {
 		"@types/pluralize": "0.0.29",
-		"@vscode/webview-ui-toolkit": "^0.8.1",
+		"@vscode/webview-ui-toolkit": "^0.8.2",
 		"pluralize": "^8.0.0"
 	}
 }
