+++ src/utilities/helpers.ts
@@ -1,0 +1,31 @@
+export function getNamespaceName(documentText: string): string{
+	const match = documentText.match(/namespace\s+(\w+)/);
+	if(match){
+		return match[1];
+	}
+	return '';
+}
+
+export function getLowerCaseClassName(className: string): string {
+	return className.charAt(0).toLowerCase() + className.slice(1);
+}
+
+export function addNamespace(usings: string[], fileNamespace: string, newNamespace: string): void {
+	if (newNamespace) {
+		// Add the new namespace if it doesn't start with the file's namespace
+		if (fileNamespace.startsWith(newNamespace)) {
+			if (newNamespace.length > fileNamespace.length) {
+				if (newNamespace[fileNamespace.length] !== '.') {
+					/* This logic handles cases like
+					*	File namespace: Project.File.Service
+					*	New  namespace: Project.File.ServiceInterface
+					*/
+					usings.push(newNamespace);
+				}
+			}
+		}
+		else {
+			usings.push(newNamespace);
+		}
+	}
+}
